requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.16.1
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
asttokens==2.0.5
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==21.5b2
celery==4.4.7
celery-pool-asyncio==0.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.3
django-celery==3.3.1
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-heroku==0.3.1
django-redis==5.0.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
drf-spectacular==0.17.0
environ==1.0
executing==0.6.0
fastapi==0.66.0
flower==0.9.7
gunicorn==20.1.0
h11==0.12.0
honcho==0.5.0
humanize==3.6.0
hyperlink==21.0.0
icecream==2.1.1
idna==2.10
incremental==21.3.0
inflection==0.5.1
iso8601==0.1.14
isort==5.8.0
jsonschema==3.2.0
kombu==4.6.11
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pylint==2.8.3
pyOpenSSL==20.0.1
pypika-tortoise==0.1.1
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
starlette==0.14.2
toml==0.10.2
tornado==6.1
tortoise-orm==0.17.4
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.14.0
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.4.0
